Change ScoreThreshold from float to double
Fixes #37

+++ AspNetCore.ReCaptcha/ReCaptchaSettings.cs
@@ -14,7 +14,7 @@ namespace AspNetCore.ReCaptcha
         public string SecretKey { get; set; }
         public ReCaptchaVersion Version { get; set; }
         public bool UseRecaptchaNet { get; set; }
-        public float ScoreThreshold { get; set; } = 0.5f;
+        public double ScoreThreshold { get; set; } = 0.5;
         public Func<Type, IStringLocalizerFactory, IStringLocalizer> LocalizerProvider { get; set; }
         public Uri RecaptchaBaseUrl { get; set; } = GoogleReCaptchaBaseUrl;
     }
